% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_control.R
\name{add_control}
\alias{add_control}
\title{Title}
\usage{
add_control(
  .map,
  control_name = c("NavigationControl", "ScaleControl", "FullscreenControl",
    "GeolocateControl", "AttributionControl"),
  control_position = c("top-left", "top-right", "bottom-left", "bottom-right"),
  ...
)
}
\arguments{
\item{.map}{A \code{\link[=maplibre]{maplibre()}} HTML widget}

\item{control_name}{The name of the control}

\item{control_position}{The position of the control.}

\item{...}{Control options depending on the type of control.}
}
\value{
The updated \code{\link[=maplibre]{maplibre()}} HTML widget
}
\description{
Title
}
