% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deck-gl.R
\name{add_deck_layers}
\alias{add_deck_layers}
\title{Add \code{deck.gl} layers to the map}
\usage{
add_deck_layers(.map, layers, tooltip = NULL)
}
\arguments{
\item{.map}{A \code{\link[=maplibre]{maplibre()}} HTML widget}

\item{layers}{A list of \code{deck.gl} layers.}

\item{tooltip}{Either a single mustache template string applied to all layers
or a dictionary where keys are layer ids and values are mustache template strings.}
}
\value{
The updated \code{\link[=maplibre]{maplibre()}} HTML widget
}
\description{
Add \code{deck.gl} layers to the map
}
\examples{
deck_grid_layer <- list(
  "@@type" = "GridLayer",
  id = "GridLayer",
  data = "https://raw.githubusercontent.com/visgl/deck.gl-data/master/website/sf-bike-parking.json",
  extruded = TRUE,
  getPosition = "@@=COORDINATES",
  getColorWeight = "@@=SPACES",
  getElevationWeight = "@@=SPACES",
  elevationScale = 4,
  cellSize = 200,
  pickable = TRUE
)

setup <- mapOptions(
  center = c(-122.4, 37.74),
  zoom = 12,
  pitch = 40,
  hash = TRUE
)

maplibre(setup, deck = TRUE) |>
  add_deck_layers(list(deck_grid_layer))
}
