% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sources.R
\name{add_source}
\alias{add_source}
\title{Add a source to map}
\usage{
add_source(.map, id, source)
}
\arguments{
\item{.map}{A \code{\link[=maplibre]{maplibre()}} HTML widget}

\item{id}{The unique ID of the source.}

\item{source}{The source to be added to the map.}
}
\value{
The updated \code{\link[=maplibre]{maplibre()}} HTML widget
}
\description{
Add a source to map
}
