% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{add_tooltip}
\alias{add_tooltip}
\title{Title}
\usage{
add_tooltip(.map, layer_id, prop)
}
\arguments{
\item{prop}{}
}
\description{
Title
}
\examples{
library(maplibre)

earthquakes_source <- list(
  type = "geojson",
  data = "https://docs.mapbox.com/mapbox-gl-js/assets/earthquakes.geojson"
)

earthquakes_layer <- Layer(
  id = "earthquakes",
  type = "circle",
  source = earthquakes_source,
  paint = list("circle-color" = "yellow", "circle-radius" = 2)
)

maplibre() |>
  add_layer(earthquakes_layer)


}
