% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemaps.R
\name{construct_basemap_style}
\alias{construct_basemap_style}
\title{Create a basemap style definition}
\usage{
construct_basemap_style(sources, layers, name = "custom-basemap")
}
\arguments{
\item{sources}{The sources to be used for the basemap style.}

\item{layers}{The layers to be used for the basemap style.}

\item{name}{The name of the basemap style.}
}
\value{
list
}
\description{
Create a basemap style definition
}
\examples{
PMTILES_URL <- "https://pmtiles.io/protomaps(vector)ODbL_firenze.pmtiles"

pmtiles_source <- list(
  type = "vector",
  url = glue::glue("pmtiles://{PMTILES_URL}"),
  attribution = '© <a href="https://openstreetmap.org">OpenStreetMap</a>'
)

sources <- list(pmtiles = pmtiles_source)

landuse <- layer(
  id = "buildings",
  source = "pmtiles",
  "source-layer" = "landuse",
  type = LayerType$FILL,
  paint = list("fill-color" = "steelblue")
)

roads <- layer(
  id = "roads",
  source = "pmtiles",
  "source-layer" = "roads",
  type = LayerType$LINE,
  paint = list("line-color" = "black"),
)

custom_basemap <- construct_basemap_style(
  sources,
  layers = list(landuse, roads)
)

setup <- map_options(
  style = custom_basemap,
  bounds = list(11.154026, 43.7270125, 11.3289395, 43.8325455)
)

maplibre(setup)
}
