% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{fullscreen_control}
\alias{fullscreen_control}
\title{A FullscreenControl control contains a button for toggling the map in and out of fullscreen mode}
\usage{
fullscreen_control(
  .map,
  position = c("top-left", "top-right", "bottom-left", "bottom-right"),
  ...
)
}
\arguments{
\item{.map}{A \code{\link[=maplibre]{maplibre()}} HTML widget}

\item{position}{The position of the control.}

\item{...}{Control options depending on the type of control.}
}
\value{
The updated \code{\link[=maplibre]{maplibre()}} HTML widget
}
\description{
A FullscreenControl control contains a button for toggling the map in and out of fullscreen mode
}
\examples{
# Add Navigation control

maplibre(map_options()) |>
  navigation_control(
    position = "top-left",
    visualize_pitch = TRUE,
    show_compass = TRUE,
    show_zoom = TRUE
  )


# Add Scale control

maplibre(map_options()) |>
  scale_control(
    position = "bottom-left",
    max_width = 200,
    unit = "metric"
  )
# Add Geolocate control

maplibre(map_options()) |>
  add_control(control_name = "GeolocateControl")



maplibre(map_options()) |>
  geolocate_control(position = "bottom-right",show_accuracy_circle = TRUE, show_user_location = TRUE,
                    track_user_location = TRUE, fit_bounds_options = list(linear = FALSE))


# Add Fullscreen control

maplibre(map_options()) |>
  add_control(control_name = "FullscreenControl",list(NULL), control_position = "bottom-right")

maplibre(map_options())|>
  fullscreen_control(position = "bottom-left")


}
