% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sources.R
\name{geojson_source}
\alias{geojson_source}
\title{GeoJSON Source}
\usage{
geojson_source(data, ...)
}
\arguments{
\item{data}{Either an URL or a GeoJSON object.}

\item{...}{Further options.}
}
\value{
list
}
\description{
GeoJSON Source
}
