% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{geolocate_control}
\alias{geolocate_control}
\title{A GeolocateControl control provides a button that uses the browser's geolocation API to locate the user on the map.}
\usage{
geolocate_control(
  .map,
  position = c("top-left", "top-right", "bottom-left", "bottom-right"),
  show_accuracy_circle = FALSE,
  show_user_location = FALSE,
  track_user_location = FALSE,
  fit_bounds_options = NULL,
  position_options = NULL
)
}
\arguments{
\item{.map}{A \code{\link[=maplibre]{maplibre()}} HTML widget}

\item{position}{The position of the control.}

\item{show_accuracy_circle}{By default, if showUserLocation is true, a transparent circle will be drawn around the user location indicating the accuracy (95\% confidence level) of the user's location. Set to FALSE to disable. Always disabled when showUserLocation is FALSE.}

\item{show_user_location}{By default a dot will be shown on the map at the user's location. Set to FALSE to disable.}

\item{track_user_location}{If TRUE the GeolocateControl becomes a toggle button and when active the map will receive updates to the user's location as it changes.}

\item{fit_bounds_options}{A options object to use when the map is panned and zoomed to the user's location. The default is to use a maxZoom of 15 to limit how far the map will zoom in for very accurate locations.}

\item{position_options}{Optional Geolocation APIs options}
}
\value{
The updated \code{\link[=maplibre]{maplibre()}} HTML widget
}
\description{
A GeolocateControl control provides a button that uses the browser's geolocation API to locate the user on the map.
}
\examples{
# Add Navigation control

maplibre(map_options()) |>
  navigation_control(
    position = "top-left",
    visualize_pitch = TRUE,
    show_compass = TRUE,
    show_zoom = TRUE
  )


# Add Scale control

maplibre(map_options()) |>
  scale_control(
    position = "top-left",
    max_width = 200,
    unit = "metric"
  )
# Add Geolocate control

maplibre(map_options()) |>
  add_control(control_name = "GeolocateControl", control_position = "top-left")


maplibre(map_options()) |>
  geolocate_control(position = "top-left",show_accuracy_circle = TRUE, show_user_location = TRUE,
                    track_user_location = TRUE, fit_bounds_options = list(linear = FALSE))
}
