% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maplibre.R
\name{mapOptions}
\alias{mapOptions}
\title{Add Map Options to a Map Object}
\usage{
mapOptions(style = basemaps$carto$dark_matter, ...)
}
\arguments{
\item{...}{}
}
\description{
Add Map Options to a Map Object
}
