% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{navigation_control}
\alias{navigation_control}
\title{Add navigation control to map}
\usage{
navigation_control(
  .map,
  position = c("top-left", "top-right", "bottom-left", "bottom-right"),
  show_compass = TRUE,
  show_zoom = TRUE,
  visualize_pitch = TRUE
)
}
\arguments{
\item{.map}{A \code{\link[=maplibre]{maplibre()}} HTML widget}

\item{position}{The position of the control.}

\item{show_compass}{If TRUE the compass button is included.}

\item{show_zoom}{If TRUE the zoom-in and zoom-out buttons are included.}

\item{visualize_pitch}{If TRUE the pitch is visualized by rotating X-axis of compass.}
}
\value{
The updated \code{\link[=maplibre]{maplibre()}} HTML widget
}
\description{
Add navigation control to map
}
\examples{
# Add Navigation control

maplibre(map_options()) |>
  navigation_control(position = "top-left",
                     visualize_pitch = TRUE,
                     show_compass = TRUE,
                     show_zoom = TRUE)
}
