% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{scale_control}
\alias{scale_control}
\title{A ScaleControl control displays the ratio of a distance on the map to the corresponding distance on the ground.}
\usage{
scale_control(
  .map,
  position = c("top-left", "top-right", "bottom-left", "bottom-right"),
  max_width = 100,
  unit = c("imperial", "metric", "nautical")
)
}
\arguments{
\item{.map}{A \code{\link[=maplibre]{maplibre()}} HTML widget}

\item{position}{The position of the control.}

\item{max_width}{The maximum length of the scale control in pixels.}

\item{unit}{Unit of the distance ('imperial', 'metric' or 'nautical').}
}
\value{
The updated \code{\link[=maplibre]{maplibre()}} HTML widget
}
\description{
A ScaleControl control displays the ratio of a distance on the map to the corresponding distance on the ground.
}
\examples{
# Add Navigation control

maplibre(map_options()) |>
  navigation_control(
    position = "top-left",
    visualize_pitch = TRUE,
    show_compass = TRUE,
    show_zoom = TRUE
  )


# Add Scale control

maplibre(map_options()) |>
  scale_control(
    position = "top-left",
    max_width = 200,
    unit = "metric"
  )
# Add Geolocate control

maplibre(map_options()) |>
  add_control(control_name = "GeolocateControl", control_position = "top-left")


maplibre(map_options()) |>
  geolocate_control(position = "top-left",show_accuracy_circle = TRUE, show_user_location = TRUE,
                    track_user_location = TRUE, fit_bounds_options = list(linear = FALSE))
}
