% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_control.R
\name{add_control}
\alias{add_control}
\title{Title}
\usage{
add_control(
  .map,
  control_name = c("NavigationControl", "ScaleControl", "FullscreenControl",
    "GeolocateControl", "AttributionControl"),
  control_position = c("top-left", "top-right", "bottom-left", "bottom-right"),
  ...
)
}
\arguments{
\item{control_name}{}
}
\description{
Title
}
