% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker.R
\name{add_marker}
\alias{add_marker}
\title{Add a marker to map}
\usage{
add_marker(.map, marker)
}
\arguments{
\item{.map}{A \code{\link[=maplibre]{maplibre()}} HTML widget}

\item{marker}{A marker object defined using \code{\link[=Marker]{Marker()}}}
}
\description{
Add a marker to map
}
\examples{
marker <- Marker(
  lng_lat = c(9.5, 51.31667),
  popup = list(text = "This is a marker", options = list(closeButton = FALSE)),
  color = "darkred"
)

maplibre(mapOptions(center = c(9.5, 51.31667)), zoom = 4) |>
  add_marker(marker)
}
