% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maplibre.R
\name{maplibre}
\alias{maplibre}
\title{Create a Map Object for MapLibre GL}
\usage{
maplibre(
  map_options = NULL,
  deck = FALSE,
  width = "100\%",
  height = NULL,
  element_id = NULL,
  ...
)
}
\arguments{
\item{map_options}{\code{\link[=map_options]{map_options()}}}

\item{deck}{Whether to add \code{deck.gl} HTML dependency.}

\item{width, height}{The width and the height of the widget.}

\item{element_id}{The unique ID of the widgets's HTML element.}

\item{...}{Further map options.}
}
\description{
Create a Map Object for MapLibre GL
}
\examples{
map_options <- map_options(
  center = c(-123.1256, 49.24658),
  hash = TRUE,
  pitch = 0,
  style = basemaps$carto$dark_matter
  )

# add control uses add_call to add a control to the map widget
maplibre(map_options, zoom = 12) |>
  add_control("NavigationControl", "top-left", showCompass = FALSE)

# this is equal to:
maplibre(map_options, zoom = 12) |>
  add_call("addControl", "NavigationControl", list(showCompass = FALSE), "top-left")
}
