% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker.R
\name{Marker}
\alias{Marker}
\title{Create a Marker}
\usage{
Marker(lng_lat, popup = NULL, ...)
}
\arguments{
\item{lng_lat}{Longitude and latitude coordinates of the marker, measured in degrees based on WGS84 (EPSG:4326) standard.}

\item{popup}{The Popup that is displayed when a user clicks on the marker.}

\item{...}{Additional marker options}
}
\description{
Create a Marker
}
\examples{
marker <- Marker(
  lng_lat = c(9.5, 51.31667),
  popup = list(text = "This is a marker", options = list(closeButton = FALSE)),
  color = "darkred"
)

maplibre(map_options(center = c(9.5, 51.31667)), zoom = 4) |>
  add_marker(marker)
}
