% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{navigation_control}
\alias{navigation_control}
\title{Add navigation control to map
A NavigationControl control contains zoom buttons and a compass.}
\usage{
navigation_control(
  .map,
  position = c("top-left", "top-right", "bottom-left", "bottom-right"),
  show_compass = TRUE,
  show_zoom = TRUE,
  visualize_pitch = TRUE
)
}
\arguments{
\item{.map}{A \code{\link[=maplibre]{maplibre()}} HTML widget}

\item{position}{The position of the control.}

\item{show_compass}{If TRUE the compass button is included.}

\item{show_zoom}{If TRUE the zoom-in and zoom-out buttons are included.}

\item{visualize_pitch}{If TRUE the pitch is visualized by rotating X-axis of compass.}
}
\value{
The updated \code{\link[=maplibre]{maplibre()}} HTML widget
}
\description{
Add navigation control to map
A NavigationControl control contains zoom buttons and a compass.
}
\examples{
# Add Navigation control

maplibre(map_options()) |>
  navigation_control(
    position = "top-left",
    visualize_pitch = TRUE,
    show_compass = TRUE,
    show_zoom = TRUE
  )


# Add Scale control

maplibre(map_options()) |>
  scale_control(
    position = "bottom-left",
    max_width = 200,
    unit = "metric"
  )
# Add Geolocate control

maplibre(map_options()) |>
  add_control(control_name = "GeolocateControl")



maplibre(map_options()) |>
  geolocate_control(position = "bottom-right",show_accuracy_circle = TRUE, show_user_location = TRUE,
                    track_user_location = TRUE, fit_bounds_options = list(linear = FALSE))


# Add Fullscreen control

maplibre(map_options()) |>
  add_control(control_name = "FullscreenControl",list(NULL), control_position = "bottom-right")

maplibre(map_options())|>
  fullscreen_control(position = "bottom-left")


}
